Implemented design patterns task

+++ src/main/java/jdbc/StudentDAO.java
@@ -1,7 +1,8 @@
-package dao;
+package jdbc;
 
 import connection.ConnectionPool;
-import domain.User;
+import dao.IStudentDAO;
+import domain.Student;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -12,12 +13,12 @@ import java.util.List;
 
 import static connection.DAOConnection.getConnectionPool;
 
-public class UserDAO implements DAO<User>{
+public class StudentDAO implements IStudentDAO {
     private ConnectionPool connectionPool = getConnectionPool();
     @Override
-    public User select(int id) {
-        String query = "SELECT id, name, surname, email, personal_id FROM users WHERE id = "  + id;
-        User user;
+    public Student select(int id) {
+        String query = "SELECT s.user_id, s.id, u.name, u.surname, u.email, u.personal_id, s.enrollment FROM users u JOIN students s on u.id = s.user_id and s.id = " + id;
+        Student student;
 
         try {
             Connection connection = connectionPool.getConnection();
@@ -25,25 +26,27 @@ public class UserDAO implements DAO<User>{
             ResultSet resultSet = statement.executeQuery();
 
             resultSet.next();
-            int userId = resultSet.getInt("id");
+            int userId = resultSet.getInt("user_id");
+            int studentId = resultSet.getInt("id");
             String name = resultSet.getString("name");
             String surname = resultSet.getString("surname");
             String email = resultSet.getString("email");
             int personalId = resultSet.getInt("personal_id");
+            int enrollment = resultSet.getInt("enrollment");
 
-            user = new User(userId, name, surname, personalId, email);
+            student = new Student(userId, name, surname, personalId, email, studentId, enrollment);
 
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-        return user;
+        return student;
     }
 
     @Override
-    public List<User> selectAll() {
-        String query = "SELECT id, name, surname, email, personal_id FROM users";
-        List<User> users = new ArrayList<>();
-        User user;
+    public List<Student> selectAll() {
+        String query = "SELECT s.user_id, s.id, u.name, u.surname, u.email, u.personal_id, s.enrollment FROM users u RIGHT JOIN students s on u.id = s.user_id";
+        List<Student> students = new ArrayList<>();
+        Student student;
 
         try {
             Connection connection = connectionPool.getConnection();
@@ -51,33 +54,33 @@ public class UserDAO implements DAO<User>{
             ResultSet resultSet = statement.executeQuery();
 
             while (resultSet.next()) {
-                int userId = resultSet.getInt("id");
+                int userId = resultSet.getInt("user_id");
+                int studentId = resultSet.getInt("id");
                 String name = resultSet.getString("name");
                 String surname = resultSet.getString("surname");
                 String email = resultSet.getString("email");
                 int personalId = resultSet.getInt("personal_id");
+                int enrollment = resultSet.getInt("enrollment");
 
-                user = new User(userId, name, surname, personalId, email);
-                users.add(user);
+                student = new Student(userId, name, surname, personalId, email, studentId, enrollment);
+                students.add(student);
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-        return users;
+        return students;
     }
 
     @Override
-    public void insert(User user) {
-        String query = "INSERT into users (name, surname, email, personal_id) VALUES (?, ?, ?, ?)";
+    public void insert(Student student) {
+        String query = "INSERT into students (user_id, enrollment) VALUES (?, ?)";
 
         try {
             Connection connection = connectionPool.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
 
-            statement.setString(1, user.getName());
-            statement.setString(2, user.getSurname());
-            statement.setString(3, user.getEmail());
-            statement.setInt(4, user.getPersonalId());
+            statement.setInt(1, student.getUserId());
+            statement.setInt(2, student.getEnrollment());
             statement.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -85,18 +88,20 @@ public class UserDAO implements DAO<User>{
     }
 
     @Override
-    public void update(User user, int id) {
-        String query = "UPDATE users SET name = ?, surname = ?, email = ?, personal_id = ? WHERE id = ?";
+    public void update(Student student, int id) {
+        String query = "UPDATE users u JOIN students s on u.id = s.user_id SET u.name = ?, u.surname = ?, u.email = ?, u.personal_id = ?, s.enrollment = ? WHERE u.id = ?";
 
         try {
             Connection connection = connectionPool.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
 
-            statement.setString(1, user.getName());
-            statement.setString(2, user.getSurname());
-            statement.setString(3, user.getEmail());
-            statement.setInt(4, user.getPersonalId() );
-            statement.setInt(5, id);
+            statement.setInt(1, student.getUserId());
+            statement.setString(2, student.getName());
+            statement.setString(3, student.getSurname());
+            statement.setString(4, student.getEmail());
+            statement.setInt(5, student.getPersonalId());
+            statement.setInt(6, student.getEnrollment());
+            statement.setInt(7, id);
             statement.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -104,14 +109,14 @@ public class UserDAO implements DAO<User>{
     }
 
     @Override
-    public void delete(User user) {
-        String query = "DELETE FROM users WHERE u.id = ?";
+    public void delete(Student student) {
+        String query = "DELETE FROM users u JOIN students s on u.id = s.user_id WHERE u.id = ?";
 
         try {
             Connection connection = connectionPool.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
 
-            statement.setInt(1, user.getUserId());
+            statement.setInt(1, student.getUserId());
             statement.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
